WIP, add manifest_store
